fix: address review remarks — rename AbstractMethodVisitor, move MethodParamsList to DTO namespace
